Fix bugs

+++ src/main/java/net/gcdc/uppertester/CamTriggerChangeSpeed.java
@@ -2,5 +2,5 @@ package net.gcdc.uppertester;
 
 public class CamTriggerChangeSpeed {
     byte messageType = 0x31;
-    byte speedVariation;  // Signed integer. Speed variation in units of m/s
+    short speedVariation;  // Signed integer. Speed variation in units of m/s
 }
